Getting started

The general workflow for using HighDimPDE.jl is as follows:

Define a Partial Integro Differential Equation problem

HighDimPDE.PIDEProblem — Type

PIDEProblem(g, f, μ, σ, x, tspan, p = nothing, x=nothing, u_domain=nothing, neumann_bc=nothing)

Defines a Partial Integro Differential Problem, of the form du/dt = 1/2 Tr(\sigma \sigma^T) Δu(t,x) + μ ∇u(t,x) + \int f(u,x) dx; where f is a nonlinear Lipschitz function

Arguments

g : The initial condition g(x, p, t).

f : The function f(x, y, u(x, t), u(y, t), ∇u(x, t), ∇u(y, t), p, t)

μ : The drift function of X from Ito's Lemma μ(x, p, t)

σ : The noise function of X from Ito's Lemma σ(x, p, t)

tspan: The timespan of the problem.

p: the parameter

x: the point of the solution required

u_domain : if provided, approximating the solution on the hypercube u_domain[1] × u_domain[2].

neumann_bc: if provided, neumann boundary conditions on the hypercube neumann_bc[1] × neumann_bc[2].

Select a solver algorithm

Solve the problem

Let's illustrate that with some examples.

MLP

Local PDE

Let's solve the Fisher KPP PDE in dimension 10 with MLP.

\[\partial_t u = u (1 - u) + \frac{1}{2}\sigma^2\Delta_xu \tag{1}\]

using HighDimPDE

## Definition of the problem
d = 10 # dimension of the problem
tspan = (0.0,0.5) # time horizon
x0 = fill(0.,d)  # initial point
g(x) = exp(- sum(x.^2) ) # initial condition
μ(x, p, t) = 0.0 # advection coefficients
σ(x, p, t) = 0.1 # diffusion coefficients
f(x, y, v_x, v_y, ∇v_x, ∇v_y, p, t) = max(0.0, v_x) * (1 -  max(0.0, v_x)) # nonlocal nonlinear part of the
prob = PIDEProblem(g, f, μ, σ, x0, tspan) # defining the problem

## Definition of the algorithm
alg = MLP() # defining the algorithm. We use the Multi Level Picard algorithm

## Solving with multiple threads 
sol = solve(prob, alg, multithreading=true)

Non local PDE with Neumann boundary conditions

Let's include in the previous equation non local competition, i.e.

\[\partial_t u = u (1 - \int_\Omega u(t,y)dy) + \frac{1}{2}\sigma^2\Delta_xu \tag{2}\]

where $\Omega = [-1/2, 1/2]^d$, and let's assume Neumann Boundary condition on $\Omega$.

using HighDimPDE

## Definition of the problem
d = 10 # dimension of the problem
tspan = (0.0,0.5) # time horizon
x0 = fill(0.,d)  # initial point
g(x) = exp( -sum(x.^2) ) # initial condition
μ(x, p, t) = 0.0 # advection coefficients
σ(x, p, t) = 0.1 # diffusion coefficients
u_domain = [-1/2, 1/2]
f(x, y, v_x, v_y, ∇v_x, ∇v_y, t) = max(0.0, v_x) * (1 -  max(0.0, v_y)) 
prob = PIDEProblem(g, f, μ, 
                    σ, x0, tspan, 
                    u_domain = u_domain) # defining u_domain is sufficient to implement Neumann boundary conditions

## Definition of the algorithm
alg = MLP(mc_sample = UniformSampling(u_domain[1], u_domain[2]) ) 

sol = solve(prob, alg, multithreading=true)

DeepSplitting

Let's solve the previous equation with DeepSplitting.

using HighDimPDE

## Definition of the problem
d = 10 # dimension of the problem
tspan = (0.0, 0.5) # time horizon
x0 = fill(0.,d)  # initial point
g(x) = exp.(- sum(x.^2, dims=1) ) # initial condition
μ(x, p, t) = 0.0 # advection coefficients
σ(x, p, t) = 0.1 # diffusion coefficients
u_domain = [-1/2, 1/2]
f(x, y, v_x, v_y, ∇v_x, ∇v_y, t) = max.(0f0, v_x) .* (1f0 .-  max.(0f0, v_y)) 
prob = PIDEProblem(g, f, μ, 
                    σ, x0, tspan, 
                    u_domain = u_domain)

## Definition of the neural network to use
using Flux # needed to define the neural network

hls = d + 50 #hidden layer size

nn = Flux.Chain(Dense(d, hls, tanh),
        Dense(hls, hls, tanh),
        Dense(hls, 1)) # neural network used by the scheme

opt = Flux.Optimiser(ExpDecay(0.1,
                0.1,
                200,
                1e-4),
                ADAM() )#optimiser

## Definition of the algorithm
alg = DeepSplitting(nn,
                    opt = opt,
                    mc_sample = UniformSampling(u_domain[1], u_domain[2]))

sol = solve(prob, 
            alg, 
            dt=0.1, 
            verbose = true, 
            abstol = 2e-3,
            maxiters = 1000,
            batch_size = 1000)

Solving on the GPU

DeepSplitting can run on the GPU for (much) improved performance. To do so, just set use_cuda = true.

sol = solve(prob, 
            alg, 
            dt=0.1, 
            verbose = true, 
            abstol = 2e-3,
            maxiters = 1000,
            batch_size = 1000,
            use_cuda=true)